refactor: refuse using deploy as a package
simplify initiating via python and make possible to define imports from project root
